Fixes input-method issue and add repl improvements (#17)
This PR adds minor revisions to the REPL and solves an issue that triggered
by typing '\'.

As it's now, the juvix input method (based on lean's) creates an object
of the class AbbreviationFeatureRewrite per document, which replaces
words with other words/symbols (using a table provided in the package
configuration). So, one rewriter object must be created every time a
Juvix file is opened, and we should be disposed of it as soon as the
document is closed. Otherwise, the rewriter creates tons of conflicts.
Before this PR, I used a function called autorun from the mobx
library for creating/disposing of some objects. It works fine in some
scenarios. However, this input method issues does not. I took
the simple vscode/ts approach, which in fact, seems to work as expected.